requirements.txt
torch>=2.0.0
transformers>=4.34.0
datasets>=2.14.0
openai-whisper>=20231117
scikit-learn>=1.3.0
tqdm>=4.66.0
numpy>=1.24.0
matplotlib>=3.7.0
soundfile>=0.12.1
